+++ lib/ts/axios.ts
@@ -26,7 +26,6 @@ import AuthHttpRequestFetch, {
     getTokenForHeaderAuth
 } from "./fetch";
 import { PROCESS_STATE, ProcessState } from "./processState";
-import { shouldDoInterceptionBasedOnUrl } from "./utils";
 import WindowHandlerReference from "./utils/windowHandler";
 import { logDebugMessage } from "./logger";
 
@@ -53,7 +52,7 @@ export async function interceptorFunctionRequestFulfilled(config: AxiosRequestCo
     try {
         doNotDoInterception =
             typeof url === "string" &&
-            !shouldDoInterceptionBasedOnUrl(
+            !AuthHttpRequestFetch.recipeImpl.shouldDoInterceptionBasedOnUrl(
                 url,
                 AuthHttpRequestFetch.config.apiDomain,
                 AuthHttpRequestFetch.config.sessionTokenBackendDomain
@@ -64,7 +63,7 @@ export async function interceptorFunctionRequestFulfilled(config: AxiosRequestCo
                 "interceptorFunctionRequestFulfilled: Trying shouldDoInterceptionBasedOnUrl with location.origin"
             );
             // .origin gives the port as well..
-            doNotDoInterception = !shouldDoInterceptionBasedOnUrl(
+            doNotDoInterception = !AuthHttpRequestFetch.recipeImpl.shouldDoInterceptionBasedOnUrl(
                 WindowHandlerReference.getReferenceOrThrow().windowHandler.location.getOrigin(),
                 AuthHttpRequestFetch.config.apiDomain,
                 AuthHttpRequestFetch.config.sessionTokenBackendDomain
@@ -155,7 +154,7 @@ export function responseInterceptor(axiosInstance: any) {
             try {
                 doNotDoInterception =
                     (typeof url === "string" &&
-                        !shouldDoInterceptionBasedOnUrl(
+                        !AuthHttpRequestFetch.recipeImpl.shouldDoInterceptionBasedOnUrl(
                             url,
                             AuthHttpRequestFetch.config.apiDomain,
                             AuthHttpRequestFetch.config.sessionTokenBackendDomain
@@ -166,7 +165,7 @@ export function responseInterceptor(axiosInstance: any) {
                     logDebugMessage("responseInterceptor: Trying shouldDoInterceptionBasedOnUrl with location.origin");
                     // .origin gives the port as well..
                     doNotDoInterception =
-                        !shouldDoInterceptionBasedOnUrl(
+                        !AuthHttpRequestFetch.recipeImpl.shouldDoInterceptionBasedOnUrl(
                             WindowHandlerReference.getReferenceOrThrow().windowHandler.location.getOrigin(),
                             AuthHttpRequestFetch.config.apiDomain,
                             AuthHttpRequestFetch.config.sessionTokenBackendDomain
@@ -304,7 +303,7 @@ export default class AuthHttpRequest {
         try {
             doNotDoInterception =
                 typeof url === "string" &&
-                !shouldDoInterceptionBasedOnUrl(
+                !AuthHttpRequestFetch.recipeImpl.shouldDoInterceptionBasedOnUrl(
                     url,
                     AuthHttpRequestFetch.config.apiDomain,
                     AuthHttpRequestFetch.config.sessionTokenBackendDomain
@@ -315,7 +314,7 @@ export default class AuthHttpRequest {
                 logDebugMessage("doRequest: Trying shouldDoInterceptionBasedOnUrl with location.origin");
                 // .origin gives the port as well..
                 doNotDoInterception =
-                    !shouldDoInterceptionBasedOnUrl(
+                    !AuthHttpRequestFetch.recipeImpl.shouldDoInterceptionBasedOnUrl(
                         WindowHandlerReference.getReferenceOrThrow().windowHandler.location.getOrigin(),
                         AuthHttpRequestFetch.config.apiDomain,
                         AuthHttpRequestFetch.config.sessionTokenBackendDomain
@@ -571,9 +570,10 @@ async function saveTokensFromHeaders(response: AxiosResponse) {
 
 async function removeAuthHeaderIfMatchesLocalToken(config: AxiosRequestConfig<any>) {
     const accessToken = await getTokenForHeaderAuth("access");
+    const refreshToken = await getTokenForHeaderAuth("refresh");
     const authHeader = config.headers!.Authorization || config.headers!.authorization;
 
-    if (accessToken !== undefined) {
+    if (accessToken !== undefined && refreshToken !== undefined) {
         if (authHeader === `Bearer ${accessToken}` || "__supertokensAddedAuthHeader" in config) {
             // We are ignoring the Authorization header set by the user in this case, because it would cause issues
             // If we do not ignore this, then this header would be used even if the request is being retried after a refresh, even though it contains an outdated access token.